BUGFIX - ApiApp.java subindo corretamente

+++ semana_4_java/src/br/com/onebox/app/domain/ProdutoIsento.java
@@ -4,7 +4,7 @@ import java.math.BigDecimal;
 
 public class ProdutoIsento extends Produto {
 
-    public ProdutoIsento(int id,
+    public ProdutoIsento(Long id,
                          String nome,
                          String descricao,
                          BigDecimal precoUnitario,
